+++ public/pages/workflow_detail/agentic_search/configure_flow/configure_flow.tsx
@@ -41,9 +41,11 @@ import {
 } from '../../../../utils';
 import { AgentConfiguration } from './agent_configuration';
 import { AgenticSearchInfoModal } from '../components';
+import '../agentic_search_styles.scss';
 
 interface ConfigureFlowProps {
   uiConfig: WorkflowConfig | undefined;
+  closePanel: () => void;
 }
 
 /**
@@ -197,40 +199,51 @@ export function ConfigureFlow(props: ConfigureFlowProps) {
           overflow: 'hidden',
         }}
       >
-        <EuiFlexItem grow={false} style={{ marginBottom: '0px' }}>
-          <EuiFlexGroup direction="row" alignItems="center" gutterSize="s">
+        <EuiFlexItem
+          grow={false}
+          style={{ marginBottom: '0px', marginLeft: '12px' }}
+        >
+          <EuiFlexGroup
+            direction="row"
+            justifyContent="spaceBetween"
+            gutterSize="s"
+          >
             <EuiFlexItem grow={false}>
-              <EuiTitle>
-                <h3>Configure agent</h3>
-              </EuiTitle>
+              <EuiFlexGroup direction="row" alignItems="center" gutterSize="s">
+                <EuiFlexItem grow={false}>
+                  <EuiTitle>
+                    <h3>Configure agent</h3>
+                  </EuiTitle>
+                </EuiFlexItem>
+                <EuiFlexItem grow={false} style={{ marginTop: '12px' }}>
+                  <EuiBetaBadge
+                    label="EXPERIMENTAL"
+                    tooltipContent="Configuring agentic search flows is an experimental feature and may change in future releases"
+                    size="s"
+                  />
+                </EuiFlexItem>
+                <EuiFlexItem
+                  grow={false}
+                  style={{ marginTop: '12px', marginLeft: '0px' }}
+                >
+                  <EuiSmallButtonIcon
+                    aria-label="What is agentic search?"
+                    iconType="questionInCircle"
+                    onClick={() => setIsModalOpen(true)}
+                  />
+                </EuiFlexItem>
+              </EuiFlexGroup>
             </EuiFlexItem>
-            <EuiFlexItem grow={false} style={{ marginTop: '12px' }}>
-              <EuiBetaBadge
-                label="EXPERIMENTAL"
-                tooltipContent="Configuring agentic search flows is an experimental feature and may change in future releases"
-                size="s"
-              />
-            </EuiFlexItem>
-            <EuiFlexItem
-              grow={false}
-              style={{ marginTop: '12px', marginLeft: '0px' }}
-            >
+            <EuiFlexItem grow={false}>
               <EuiSmallButtonIcon
-                aria-label="What is agentic search?"
-                iconType="questionInCircle"
-                onClick={() => setIsModalOpen(true)}
+                iconType="menuLeft"
+                onClick={() => props.closePanel()}
+                aria-label="closeAgentConfig"
               />
             </EuiFlexItem>
           </EuiFlexGroup>
         </EuiFlexItem>
-        <EuiFlexItem
-          style={{
-            overflowY: 'auto',
-            scrollbarGutter: 'auto',
-            scrollbarWidth: 'auto',
-            overflowX: 'hidden',
-          }}
-        >
+        <EuiFlexItem className="agentic-search-workspace-panel">
           <EuiPanel paddingSize="none" hasBorder={false} hasShadow={false}>
             <EuiFlexGroup direction="column">
               <EuiFlexItem>